Start work with vector

+++ src/restl/new.hpp
@@ -1,9 +1,7 @@
-#include <cstddef>
-
 #ifndef NEW_H_
 #define NEW_H_
 
-inline void* operator new(size_t, void* ptr) {
+inline void* operator new(unsigned int, void* ptr) {
     return ptr;
 }
